# Refactor checkandEstimate to optimize podReconciliation