For the function `checkandEstimate`. I just test it using this case.

<details>
  <summary>test case</summary>

```
 apiVersion: argoproj.io/v1alpha1
kind: Workflow
metadata:
  generateName: pod-limits-
spec:
  entrypoint: pod-limits
  arguments:
    parameters:
    - name: limit
      value: 300

  templates:
  - name: pod-limits
    steps:
    - - name: run-pod
        template: run-pod
        withSequence:
          count: "{{workflow.parameters.limit}}"
    - - name: run-pod-2
        template: run-pod
        withSequence:
          count: "{{workflow.parameters.limit}}"
    - - name: run-pod-3
        template: run-pod
        withSequence:
          count: "{{workflow.parameters.limit}}"
    - - name: run-pod-4
        template: run-pod
        withSequence:
          count: "{{workflow.parameters.limit}}"
    - - name: run-pod-5
        template: run-pod
        withSequence:
          count: "{{workflow.parameters.limit}}"
    - - name: run-pod-6
        template: run-pod
        withSequence:
          count: "{{workflow.parameters.limit}}"
    - - name: run-pod-7
        template: run-pod
        withSequence:
          count: "{{workflow.parameters.limit}}"
    - - name: run-pod-8
        template: run-pod
        withSequence:
          count: "{{workflow.parameters.limit}}"
    - - name: run-pod-9
        template: run-pod
        withSequence:
          count: "{{workflow.parameters.limit}}"
    - - name: run-pod-10
        template: run-pod
        withSequence:
          count: "{{workflow.parameters.limit}}"

  - name: run-pod
    container:
      image: "alpine:3.7"
      command: [sh, -c]
      args: ["echo sleeping 1s; sleep 1"]
```

</details>


From the workflow pod log, we can see that, it will reach to context deadline.

```
time="2019-04-08T03:41:15Z" level=warning msg="Deadline exceeded" namespace=default workflow=pod-limits-jbzx6
time="2019-04-08T03:40:45Z" level=info msg="podReconciliation list pod begin 2019-04-08 03:40:45.068890985 " namespace=default workflow=pod-limits-jbzx6
time="2019-04-08T03:40:54Z" level=info msg="podReconciliation list pod end 2019-04-08 03:40:54.658723708" namespace=default workflow=pod-limits-jbzx6
```

Every pod list action will cost about 10s.

After using this PR,  the time will reduce to 1s, the log can be seen here.

```
time="2019-04-08T04:05:18Z" level=info msg="podReconciliation list pod begin 2019-04-08 04:05:18.767515515" namespace=default workflow=pod-limits-57wpl
time="2019-04-08T04:05:18Z" level=info msg="podReconciliation list pod end 2019-04-08 04:05:18.787992764 " namespace=default workflow=pod-limits-57wpl
```

List pod action use about 1s.


I think we shouldn't judge whether each node exceeds the max size limit.   We can just judge it for every whole workflow.  If the workflow exceeds the max size limit, we can just send the error message and mark the workflow error, back the status to the before.

This pr can reduce much time waste in size limit judge, and when size exceeds the max limit,  the workflow will also be marked error.   

Pls @sarabala1979 @jessesuen take a look at it.     
It will be very helpful for 2000-3000+ steps workflow.