@sarabala1979 2019-04-08
My only concern about this implementation is if each step produces a different size of outputs, Workflow will fail with the biggest output node and ignore rest of nodes which may fit into size.
The above implementation will work only if all nodes have the same size output.  

Let me talk to @jessesuen and come up with a proposal